pyproject.toml
[project]
name = "dishlingo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
    "aiogram>=3.21.0",
    "requests>=2.32.4",
    "google-generativeai"
]
